fix discussions
add annotation
