--- src/main/resources/META-INF/defaultservices/com.codeborne.selenide.Clipboard
@@ -1,1 +1,0 @@
-com.codeborne.selenide.Clipboard
