+++ src/main/resources/META-INF/defaultservices/com.codeborne.selenide.ClipboardService
@@ -1,0 +1,1 @@
+com.codeborne.selenide.ClipboardService
